+++ packages/migrate/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @contentmap/migrate
 
+## 0.5.0
+
 ## 0.4.1
 
 ## 0.4.0